--- ranker/utils/file2.py
@@ -1,4 +1,0 @@
-from hmm import getYtUrl
-
-x =  getYtUrl("Blinding lights")
-print(x)
