+++ requirements.txt
@@ -39,6 +39,6 @@ virtualenv==20.0.17
 virtualenv-clone==0.5.4
 whitenoise==5.0.1
 wrapt==1.12.1
-youtube-dl==2020.5.3
+youtube-dl==2020.12.14
 zipp==3.1.0
 selenium==3.141.0
